cvorb/test: remove warning
--remove following warning:
cvorbtest.c: In function ‘h_pcb’:
cvorbtest.c:293: warning: format ‘%016llx’ expects type ‘long long
unsigned int’, but argument 2 has type ‘uint64_t’

+++ cvorb/test/cvorbtest.c
@@ -290,7 +290,7 @@ int h_pcb(struct cmd_desc *cmdd, struct atom *atoms)
 		return 1;
 	}
 	cvorb_get_pcb_id(current_dev, &pcb_id);
-	printf ("0x%016llx\n", pcb_id);
+	printf ("0x%016llx\n", (long long unsigned int) pcb_id);
 	return 1;
 }
